Fix triliovault_enable_tls_backend variable
Backend TLS is currently broken for Trilio since the scripts do not
deploy the proxy containers that perform the TLS termination. This
corrects a condition in the template which was using a variable meant
for glance (presumably as this role was once based on glance). This at
least allows you to set `triliovault_enable_tls_backend: false`